LayoutTests/editing/text-iterator/sequential-collapsed-ranges-expected.txt

a b

Testing first letter with narrow width...
  Default behavior
    'a' (location 0 length 1)
  IgnoresWhiteSpaceAtEndOfRun
    'a' (location 0 length 1)

Testing first letter with wide width...
  Default behavior
    'a' (location 0 length 1)
  IgnoresWhiteSpaceAtEndOfRun
    'a' (location 0 length 1)

Testing everything except last letter with narrow width...
  Default behavior
    'a' (location 0 length 1)
    ' ' (location 1 length 0)
  IgnoresWhiteSpaceAtEndOfRun
    'a' (location 0 length 1)

Testing everything except last letter with wide width...
  Default behavior
    'a ' (location 0 length 2)
  IgnoresWhiteSpaceAtEndOfRun
    'a ' (location 0 length 2)

Testing everything with narrow width...
  Default behavior
    'a' (location 0 length 1)
    ' ' (location 1 length 0)
    'b' (location 2 length 1)
    '\n' (location 3 length 0)
  IgnoresWhiteSpaceAtEndOfRun
    'a' (location 0 length 1)
    ' ' (location 1 length 0)
    'b' (location 1 length 1)
    '\n' (location 3 length 0)

Testing everything with wide width...
  Default behavior
    'a ' (location 0 length 2)
    'b' (location 2 length 1)
    '\n' (location 3 length 0)
  IgnoresWhiteSpaceAtEndOfRun
    'a ' (location 0 length 2)
    'b' (location 2 length 1)
    '\n' (location 3 length 0)

PASS successfullyParsed is true

TEST COMPLETE
